Get rid of SaitekAviatorStick.cs
