+++ MobiFlight/Joysticks/LabeledJoystick.cs
@@ -7,8 +7,13 @@ namespace MobiFlight.Joysticks
     public class LabeledJoystick : Joystick
     {
         protected Dictionary<string, string> Labels = new Dictionary<string, string>();
+        readonly private HidDefinition Definition;
 
-        public LabeledJoystick(SharpDX.DirectInput.Joystick joystick) : base(joystick) { }
+        public LabeledJoystick(SharpDX.DirectInput.Joystick joystick, HidDefinition definition) : base(joystick) 
+        {
+            Definition = definition;
+            Definition?.Inputs.ForEach(input => Labels.Add(input.Name, input.Label));
+        }
 
         protected override List<JoystickDevice> GetButtonsSorted()
         {
@@ -16,6 +21,12 @@ namespace MobiFlight.Joysticks
             buttons.Sort(SortByPositionInDictionary);
 
             return buttons;
+        }
+
+        protected override void EnumerateOutputDevices()
+        {
+            base.EnumerateOutputDevices();
+            Definition?.Outputs.ForEach(output => Lights.Add(new JoystickOutputDevice() { Label = output.Label, Name = output.Name, Byte = output.Byte, Bit = output.Bit }));
         }
 
         public override string MapDeviceNameToLabel(string name)
